# ENH: Fast paths for richcompare