--- a/numpy/core/src/multiarray/scalartypes.c.src
+++ b/numpy/core/src/multiarray/scalartypes.c.src
@@ -4144,7 +4144,7 @@ is_anyscalar_exact(PyObject *obj)
 }
 
 NPY_NO_EXPORT int
-_typenum_fromtypeobj(PyObject *type, int user)
+PyArray_TypeNumFromNumPyScalarType(PyObject *type, int user)
 {
     int typenum, i;