--- a/numpy/core/src/multiarray/scalartypes.h
+++ b/numpy/core/src/multiarray/scalartypes.h
@@ -26,7 +26,7 @@ NPY_NO_EXPORT int
 is_anyscalar_exact(PyObject *obj);
 
 NPY_NO_EXPORT int
-_typenum_fromtypeobj(PyObject *type, int user);
+PyArray_TypeNumFromNumPyScalarType(PyObject *type, int user);
 
 NPY_NO_EXPORT void *
 scalar_value(PyObject *scalar, PyArray_Descr *descr);