--- a/numpy/core/src/umath/scalarmath.c.src
+++ b/numpy/core/src/umath/scalarmath.c.src
@@ -15,6 +15,7 @@
 #include "numpy/arrayobject.h"
 #include "numpy/ufuncobject.h"
 #include "numpy/arrayscalars.h"
+
 #include "scalartypes.h"
 
 #include "npy_import.h"
@@ -595,10 +596,8 @@ descr_from_basic_scalar(PyObject *obj, PyArray_Descr **descr)
         return 1;
     }
     else if (PyLong_CheckExact(obj)) {
-        int overflow;
-        if (PyLong_AsLongAndOverflow(obj, &overflow) < 0) {
-            /* overflow must have occurred */
-            assert(overflow != 0);
+        if (PyLong_AsLong(obj) == -1 && PyErr_Occurred()) {
+            PyErr_Clear();
             return -1;
         }
         *descr = PyArray_DescrFromType(NPY_LONG);
@@ -625,33 +624,37 @@ descr_from_basic_scalar(PyObject *obj, PyArray_Descr **descr)
  */
 static PyObject*
 do_richcompare_on_scalars(PyObject *self, PyObject *other, int cmp_op) {
-    PyObject *item_self, *item_other, *ret;
-    PyArray_Descr *self_descr, *other_descr;
-    void *data_self, *data_other;
-    int pyscalar_self, pyscalar_other;
+    PyObject *item_self, *ret=NULL;
+    PyArray_Descr *self_descr=NULL, *other_descr=NULL;
+    void *data_self;
+    int pyscalar_other, pyscalar_self;
+    int is_complex_operands, is_equality_operator;
 
     pyscalar_self = descr_from_basic_scalar(self, &self_descr);
     pyscalar_other = descr_from_basic_scalar(other, &other_descr);
 
-    if (NPY_LIKELY(pyscalar_self > 0 && pyscalar_other > 0)) {
-        data_self = scalar_value(self, NULL);
-        data_other = scalar_value(other, NULL);
+    if (pyscalar_self >= 0 && pyscalar_other >= 0) {
+        /*
+         * If either of the operands are complex AND the operator is not equality,
+         * python's built-in richcompare cannot be used as it is not supported
+         */
+        is_complex_operands = (PyTypeNum_ISCOMPLEX(self_descr->type_num) ||
+            PyTypeNum_ISCOMPLEX(other_descr->type_num));
+        is_equality_operator = (cmp_op != Py_EQ && cmp_op != Py_NE);
 
-        item_self = self_descr->f->getitem(data_self, NULL);
-        item_other = other_descr->f->getitem(data_other, NULL);
-
-        Py_DECREF(item_other);
-        Py_DECREF(item_self);
-
-        ret = PyObject_RichCompare(item_self, item_other, cmp_op);
-
-        Py_DECREF(self_descr);
-        Py_DECREF(other_descr);
-
-        return ret;
+        if (!(is_complex_operands && is_equality_operator)) {
+            data_self = scalar_value(self, NULL);
+            item_self = self_descr->f->getitem(data_self, NULL);
+            if (item_self != NULL) {
+                ret = PyObject_RichCompare(item_self, other, cmp_op);
+            }
+            Py_XDECREF(item_self);
+        }
     }
+    Py_XDECREF(self_descr);
+    Py_XDECREF(other_descr);
 
-    return NULL;
+    return ret;
 }