## Changes:
1. Added check to see if RHS is a python number and prevent the creation of array.
2. Added bench

## Benchmarks:
<details>
<summary>Benchmarks:</summary>

<pre>                                                                                                                   
********************************************************************************                                                                          
WARNING: you have uncommitted changes --- these will NOT be benchmarked!                                                                                  
********************************************************************************                                                                          
· Creating environments                                                      
· Discovering benchmarks                                                     
·· Uninstalling from virtualenv-py3.8-Cython                                 
·· Building 933620b4 <enh_14415_fast_compare> for virtualenv-py3.8-Cython..................................                                                                                                                                                                                                          
·· Installing 933620b4 <enh_14415_fast_compare> into virtualenv-py3.8-Cython.                                                                             
· Running 2 total benchmarks (2 commits * 1 environments * 1 benchmarks)                                                                                  
[  0.00%] · For numpy commit 4d290795 <master> (round 1/2):                                                                                               
[  0.00%] ·· Building for virtualenv-py3.8-Cython.                                                                                                                                                                                                                                                                   
[  0.00%] ·· Benchmarking virtualenv-py3.8-Cython                                                                                                         
[ 25.00%] ··· Running (bench_ufunc.CustomScalarEqual.time_scalar_equality--).                                                                             
[ 25.00%] · For numpy commit 933620b4 <enh_14415_fast_compare> (round 1/2):                                                                               
[ 25.00%] ·· Building for virtualenv-py3.8-Cython.                                                                                                        
[ 25.00%] ·· Benchmarking virtualenv-py3.8-Cython                                                                                                         
[ 50.00%] ··· Running (bench_ufunc.CustomScalarEqual.time_scalar_equality--).                                                                             
[ 50.00%] · For numpy commit 933620b4 <enh_14415_fast_compare> (round 2/2):                                                                               
[ 50.00%] ·· Benchmarking virtualenv-py3.8-Cython                                                                                                         
[ 75.00%] ··· bench_ufunc.CustomScalarEqual.time_scalar_equality                                                                                                              ok                                                                                                                                     
[ 75.00%] ··· =============== ==========                                     
                   dtype                                                                                                                                  
              --------------- ----------                                     
                 numpy.int8    439±6ms                                       
                numpy.int16    445±8ms                                       
                numpy.int32    435±10ms                                      
                numpy.int64    147±2ms                                       
                numpy.uint8    427±10ms                                      
                numpy.uint16   444±8ms                                       
                numpy.uint32   437±10ms                                      
                numpy.uint64   447±4ms                                       
               numpy.float32   471±10ms                                      
               numpy.float64   180±4ms                                                                                                                    
              =============== ==========                                                                                                                  
                                                                             
[ 75.00%] · For numpy commit 4d290795 <master> (round 2/2):                                                                                               
[ 75.00%] ·· Building for virtualenv-py3.8-Cython.                                                                                                        
[ 75.00%] ·· Benchmarking virtualenv-py3.8-Cython                                                                                                         
[100.00%] ··· bench_ufunc.CustomScalarEqual.time_scalar_equality                                                                                                              ok                                                                                                                                     
[100.00%] ··· =============== ============                                   
                   dtype                                                     
              --------------- ------------                                   
                 numpy.int8     1.35±0s                                      
                numpy.int16    1.34±0.01s                                    
                numpy.int32    1.34±0.02s                                    
                numpy.int64     147±3ms                                      
                numpy.uint8     1.34±0s                                      
                numpy.uint16    1.36±0s                                      
                numpy.uint32    1.35±0s                                      
                numpy.uint64   1.54±0.01s                                    
               numpy.float32    1.54±0s                                      
               numpy.float64    184±4ms                                      
              =============== ============                                   

       before           after         ratio                                  
     [4d290795]       [933620b4]                                             
     <master>         <enh_14415_fast_compare>                                                                                                            
-      1.34±0.01s          445±8ms     0.33  bench_ufunc.CustomScalarEqual.time_scalar_equality(<class 'numpy.int16'>)                                                                                                                                                                                               
-         1.36±0s          444±8ms     0.33  bench_ufunc.CustomScalarEqual.time_scalar_equality(<class 'numpy.uint16'>)                                                                                                                                                                                              
-      1.34±0.02s         435±10ms     0.33  bench_ufunc.CustomScalarEqual.time_scalar_equality(<class 'numpy.int32'>)                                                                                                                                                                                               
-         1.35±0s          439±6ms     0.33  bench_ufunc.CustomScalarEqual.time_scalar_equality(<class 'numpy.int8'>)                                                                                                                                                                                                
-         1.35±0s         437±10ms     0.32  bench_ufunc.CustomScalarEqual.time_scalar_equality(<class 'numpy.uint32'>)                                                                                                                                                                                              
-         1.34±0s         427±10ms     0.32  bench_ufunc.CustomScalarEqual.time_scalar_equality(<class 'numpy.uint8'>)                                                                                                                                                                                               
-         1.54±0s         471±10ms     0.31  bench_ufunc.CustomScalarEqual.time_scalar_equality(<class 'numpy.float32'>)                                                                                                                                                                                             
-      1.54±0.01s          447±4ms     0.29  bench_ufunc.CustomScalarEqual.time_scalar_equality(<class 'numpy.uint64'>)                                                                                                                                                                                              

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.                                  
PERFORMANCE INCREASED.                                                       

</pre>
</details>

resolves: #14415

cc: @seberg , @eric-wieser , @mattip