@seberg 2020-12-09
Thanks, I will have to think about it more. I do guess that it might be nice to intercept some of this even earlier, but I do not mind injecting a fast-path there it is practical and doesn't add much complexity.

One thing to think about `.item()` has two awkward cases:

1. Low precision floating point numbers are upcast to float64, I expect that is fine, but did not think it through.
2. High precision floating point numbers (longdouble, clongdouble) are *unchanged*, that means the fast-path is definitely broken for them. Maybe that is the test-failure, maybe not.

(To be fair, the 2. part is already broken in some similar paths, but this would break it more/worse.)