@ganesh-k13 2020-12-10
I tried something like this:
```c
* #Name = Long*11, Float*3,,,#
* #fastpath = 1*14, 0*3#

.....

   case -2:
        /* use ufunc */
        if (PyErr_Occurred()) {
            return NULL;
        }
#if @fastpath@
        if (PyArray_IsPythonNumber(other)) {
            PyObject *scalar, *result;
            scalar = Py@Name@_From@Name@(arg1);
            result = PyObject_RichCompare(scalar, other, cmp_op);
            if (result == NULL) {
                return NULL;
            }
            out = PyObject_IsTrue(result);
            if (out == -1) {
                return NULL;
            }
            Py_DECREF(result);
            goto done;
        }
#endif
        return PyGenericArrType_Type.tp_richcompare(self, other, cmp_op);
```
But it fails for `np.int64(-1) == 18446744073709551615` because arg1 becomes `-1`. Even if I replace the code with `.item` I get same, anything I can do differently here?