@seberg 2020-12-10
Ah, right, I am not sure you need the `* #fastpath = 1*14, 0*3#` addition, because longdouble takes the `-3` case?

Why are you converting the `obj` to the scalar instead of the other way around? Your old code with `item = gentype_generic_method(self, NULL, NULL, "item");` made more sense to me (albeit it should be replaced with the correct C-side incarnation).

It all looks weird, but when `other` is an arbitrary Python object, and the numpy scalar is not longdouble/clongdouble, we actually currently already do:

1. Call ufuncs
2. Convert both into an array
    * The second object will be an `object` dtype array
3. Cast the original scalar (now an array) to `object` so that it has the same type as the other
    * This cast, for the main builtin numerical types, actually does `arr.item()` and not just `arr[()]`.

So in the end we already operate on Python objects, just in a very slow and convoluted way.