@ganesh-k13 2020-12-10
Oh I see, yeah I missed the `_convert2_to_ctypes` logic. `gentype_generic_method` was static to the file, hence needed conversion to a scalar to call its richcompare, hence the obj to scalar. 

Given that the original commit is showing significant gains in performance, I'll dig in a bit to see where we are slowing down. 

> albeit it should be replaced with the correct C-side incarnation

Yeah, probably moving previous code up a function call should do it, will do so after reading a bit.