@seberg 2020-12-10
Ohh, I think you are testing mixed types? I.e. one of them is an arbitrary type, the other is always a default integer (probably `C` long?)...  I.e. `int8 + int64` is *super* slow, but `int64 + int8` is fine.

IIRC those paths were messed up, because we did not have a correct "hierarchy" in the code. I tried to clean it up, but it turned out a pretty big PR and I never opened it, I should probably reactivate. There was a problem in that what we should be doing is:
```
if other.dtype.type_num > self.dtype.type_num:
    return NotImplemented
```
(At least that seemed like the easiest solution).  Right now I think what happens it that if the first dtype is the smaller type, we will always end up in the super-slow path. With the `return NotImplemented` fix, we at least only end up in the super-slow path if `(a + b).dtype is not in (a.dtype, b.dtype)`, which is much better.

More complex or generic schemes seemed like over-engineering to me, as they would have to duplicate a lot of ufunc logic which is likely not going to end up very light-weight in any case.

I would love if if you can find a more minimal fix, but in many ways I think this code requires an overhaul :(. Which means rewriting a bunch of it.