@ganesh-k13 2020-12-10
I started with your [branch](https://github.com/seberg/numpy/tree/scalar-binop-giveup-in-fallback) and resolved the conflicts(not logical conflicts, as it does not compile yet :) ) here: https://github.com/ganesh-k13/numpy/pull/1. 

I'll be happy to resume work on it and finish it if you feel it helps this part of the code.