@ganesh-k13 2020-12-12
Hey @seberg , I leveraged some stuff from the branch and added fastpath in the `-1` case where it's safe to cast. Let me know your thoughts on this.