@ganesh-k13 2020-12-16
```
       before           after         ratio
     [4d290795]       [3115c165]
     <master>         <enh_14415_fast_compare>
-      23.7±0.1μs      11.7±0.04μs     0.49  bench_scalar.ScalarMath.time_compare('int32')
-      23.9±0.3μs       11.7±0.2μs     0.49  bench_scalar.ScalarMath.time_compare('int16')
-      25.5±0.3μs       12.4±0.2μs     0.49  bench_scalar.ScalarMath.time_compare('complex64')
-     25.5±0.05μs      11.9±0.09μs     0.47  bench_scalar.ScalarMath.time_compare('float32')
-      26.1±0.4μs      11.9±0.05μs     0.45  bench_scalar.ScalarMath.time_compare('float16')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.

```