@ganesh-k13 2021-01-11
Hi @seberg , sorry for the long pause in the PR.
I noticed the following TC causing crash in specific environment. I think handling return value for `self_descr` for overflow should take care of it. 
Now to the overflow: https://github.com/numpy/numpy/blob/a190258d4e90f2a17a9469e5dd9fb5f4b045aa90/numpy/core/tests/test_getlimits.py#L68-L70
Is it defined behaviour to do `np.uint(-1)`?