@ganesh-k13 2021-01-18
Hey @seberg , I tried multiple things to reproduce the issue locally, tried `quay.io/pypa/manylinux2010_i686` and used a native 32-bit system as well, but it's not crashing. Any idea why that particular TC is failing? I followed the memory through the code, no double free or invalid dereference either.