@ganesh-k13 2021-01-18
Also not sure if I am setting `quay.io/pypa/manylinux2010_i686` correctly, for instance, I don't see python being installed in the image, yet it's there. I was curious about how are the packages being installed in the image.