@seberg 2021-01-18
Maybe that overflow `assert` is simply incorrect? It does seem a bit off.  Otherwise, my best idea is to install `gdb` in the CI and the job with that. Hopefully, getting a traceback will be sufficient. I did something for example [here (the command may be useful)](https://github.com/numpy/numpy/pull/17750/files) (but for the travis CI).