@ganesh-k13 2021-01-19
Benchmarks after the changes:
```
       before           after         ratio
     [f5f845b7]       [fe3a8975]
     <master>         <enh_14415_fast_compare>
+         583±4ns          622±4ns     1.07  bench_scalar.ScalarMath.time_addition('longfloat')
-     3.63±0.03μs      3.45±0.03μs     0.95  bench_scalar.ScalarMath.time_addition_pyint('float32')
-       584±0.8ns          554±2ns     0.95  bench_scalar.ScalarMath.time_multiplication('complex256')
-         563±7ns          531±2ns     0.94  bench_scalar.ScalarMath.time_addition('float32')
-         567±2ns          527±3ns     0.93  bench_scalar.ScalarMath.time_multiplication('int64')
-      23.7±0.2μs      11.5±0.07μs     0.49  bench_scalar.ScalarMath.time_compare('int32')
-      24.1±0.7μs       11.7±0.2μs     0.48  bench_scalar.ScalarMath.time_compare('int16')
-      25.9±0.3μs      12.1±0.05μs     0.47  bench_scalar.ScalarMath.time_compare('complex64')
-      26.2±0.2μs      12.0±0.07μs     0.46  bench_scalar.ScalarMath.time_compare('float32')
-      26.3±0.5μs      11.9±0.03μs     0.45  bench_scalar.ScalarMath.time_compare('float16')
```